+++ src/main/kotlin/blackjack/Blackjack.kt
@@ -1,0 +1,3 @@
+package blackjack
+
+class Blackjack
