+++ src/main/kotlin/blackjack/model/Suit.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.card
+package blackjack.model
 
 enum class Suit {
     Spade, Diamond, Heart, Clover
